@@ -14,14 +14,14 @@
     </parent>
 
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <spring-boot.version>2.7.15</spring-boot.version>
-        <spring.cloud.version>3.1.2</spring.cloud.version>
-        <spring.consul.version>3.1.2</spring.consul.version>
-        <spring.mybatis.version>2.3.1</spring.mybatis.version>
-        <alibaba.druid.version>1.2.18</alibaba.druid.version>
+        <spring-boot.version>3.1.3</spring-boot.version>
+        <spring.cloud.version>4.0.3</spring.cloud.version>
+        <spring.consul.version>4.0.3</spring.consul.version>
+        <spring.mybatis.version>3.0.2</spring.mybatis.version>
+        <alibaba.druid.version>1.2.19</alibaba.druid.version>
     </properties>
 
     <!--开源协议...-->
