@@ -33,22 +33,22 @@
 
     <properties>
         <!--JDK版本-->
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <!--源代码使用的JDK版本-->
-        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.source>17</maven.compiler.source>
         <!--需要生成的目标class文件的编译版本-->
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.target>17</maven.compiler.target>
         <!--maven项目源代码编码格式，参考：https://maven.apache.org/plugins/maven-resources-plugin/examples/encoding.html-->
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <commons-io.version>2.11.0</commons-io.version>
-        <commons-lang3.version>3.13.0</commons-lang3.version>
+        <commons-lang3.version>3.12.0</commons-lang3.version>
         <google-guava.version>32.1.0-jre</google-guava.version>
         <alibaba-ttl.version>2.14.3</alibaba-ttl.version>
         <jackson-databind.version>2.13.5</jackson-databind.version>
         <junit-test.version>4.13.2</junit-test.version>
         <java-jwt.version>4.4.0</java-jwt.version>
         <bcprov-jdk15on.version>1.70</bcprov-jdk15on.version>
-        <logback.version>1.2.12</logback.version>
+        <logback.version>1.4.11</logback.version>
         <jansi.version>2.4.0</jansi.version>
     </properties>
 
