@@ -2,12 +2,12 @@ package com.emily.infrastructure.test.controller;
 
 import com.emily.infrastructure.captcha.Captcha;
 import com.emily.infrastructure.captcha.CaptchaUtils;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 /**
