@@ -1,6 +1,9 @@
 package com.emily.infrastructure.test.controller;
 
 import com.emily.infrastructure.rabbitmq.factory.RabbitMqFactory;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -13,8 +16,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("api/rabbit")
 public class RabbitController {
+    @Autowired
+    private RabbitTemplate rabbitTemplate;
+    @Autowired
+    @Qualifier("emilyRabbitTemplate")
+    private RabbitTemplate rabbitTemplateEmily;
     @GetMapping("send")
     public void send(){
-        RabbitMqFactory.getRabbitTemplate("test").convertAndSend("emily.test", "", "测试故障恢复");
+        rabbitTemplate.convertAndSend("emily.test", "", "测试故障恢复");
+        rabbitTemplateEmily.convertAndSend("exchange_emily", "exchange.#","测试故障恢复");
+        RabbitMqFactory.getRabbitTemplate("emily").convertAndSend("exchange_emily", "exchange.#", "测试故障恢复");
     }
 }
