@@ -1,11 +1,11 @@
 package com.emily.cloud.test.api.websocket;
 
 import com.emily.infrastructure.logger.LoggerFactory;
+import jakarta.websocket.*;
+import jakarta.websocket.server.ServerEndpoint;
 import org.slf4j.Logger;
 import org.springframework.stereotype.Component;
 
-import javax.websocket.*;
-import javax.websocket.server.ServerEndpoint;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
