@@ -38,7 +38,7 @@ public class RabbitDelaySender {
                 //更新数据库，可靠性投递机制
                 System.out.println("正常ack-" + ack + ",id-" + correlationData.getId());
                 try {
-                    System.out.println(new String(correlationData.getReturnedMessage().getBody()));
+                    System.out.println(new String(correlationData.getReturned().getMessage().getBody()));
 
                 } catch (Exception e) {
 
@@ -49,28 +49,10 @@ public class RabbitDelaySender {
     /**
      * 创建一个消息是否被队列接收的监听对象，如果没有队列接收发送出的消息，则调用此方法进行后续处理
      */
-    private final RabbitTemplate.ReturnCallback returnCallback = new RabbitTemplate.ReturnCallback() {
-        /**
-         *
-         * @param message 被退回的消息
-         * @param replyCode 错误编码
-         * @param replyText 错误描述
-         * @param exchange 交换器
-         * @param routingKey 路由
-         */
+    private final RabbitTemplate.ReturnsCallback returnCallback = new RabbitTemplate.ReturnsCallback() {
         @Override
-        public void returnedMessage(Message message, int replyCode, String replyText, String exchange, String routingKey) {
-            System.err.println("spring_returned_message_correlation:" + message.getMessageProperties().getHeaders().get(PublisherCallbackChannel.RETURNED_MESSAGE_CORRELATION_KEY)
-                    + "return exchange: " + exchange
-                    + ", routingKey: " + routingKey
-                    + ", replyCode: " + replyCode
-                    + ", replyText: " + replyText
-                    + ",message:" + message);
-            try {
-                System.out.println(new String(message.getBody()));
-            } catch (Exception e) {
+        public void returnedMessage(ReturnedMessage returnedMessage) {
 
-            }
         }
     };
     /**
@@ -117,7 +99,7 @@ public class RabbitDelaySender {
         /**
          * 设置消息退回回调函数
          */
-        this.rabbitTemplate.setReturnCallback(returnCallback);
+        this.rabbitTemplate.setReturnsCallback(returnCallback);
         /**
          * 新增消息转换完成后、发送之前的扩展点
          */
@@ -151,7 +133,7 @@ public class RabbitDelaySender {
             /**
              * 设置此相关数据的返回消息
              */
-            correlationData.setReturnedMessage(returnedMessage);
+            //correlationData.setReturned(returnedMessage);
             /**
              * 如果msg是org.springframework.amqp.core.Message对象的实例，则直接返回，否则转化为Message对象
              */
